// src/types/car.ts
export type CarParts = {
  engine: number
  gearbox: number
  chassis: number
  frontWing: number
  rearWing: number
  suspension: number
  brakes: number
  cooling: number
  floor: number
}

export type CarStats = {
  power: number
  acceleration: number
  grip: number
  tyreWear: number
  reliability: number
  balance: number
}